fix: 잘못된 WinningLotto 역할 Lotto로 이동
