+++ docs/README.md
@@ -44,10 +44,9 @@
 > -[ ] 에러메시지를 전달받아 출력
    >   - [ ] 모든 에러 문구 앞에 [ERROR] 붙이기
 
-## YieldCalculator
-> ### 수익률 계산하는 곳
-> -[ ] 등수별 금액, 총 당첨금, 구매 금액으로 생성
-> -[ ] 일치한 금액권 * 개수로 총 당첨금을 계산
+## MoneyManagement
+> ### 구매 수량과 수익률을 계산하는 곳
+> -[ ] 구매 금액으로 수량을 계산
 > -[ ] 당첨금/구매 금액으로 수익률 계산
 > -[ ] 수익률은 소수점 둘째 자리에서 반올림
 
@@ -63,18 +62,18 @@
 > ### 등수별 메시지, 당첨 금액, 당첨 개수로 구성
 
 ## Lotto
-> ### 로또 번호 6개와 와 보너스 번호로 구성
+> ### 로또 번호 6개로 구성
 > 
 
 ## Number
 > ### 로또 번호의 각 번호 1개
 
 ## WinningLotto
-> ### 로또 당첨 번호
+> ### 사용자가 정한 로또 당첨 번호
+> ### 로또번호 6개와 보너스번호로 구성
 
 ## LottoController
 > ### OutputView와 InputView를 주입받아 생성
-> ### 
 
 ## MoneyManagement
 > ### 로또 구매 금액을 관리
