+++ src/main/java/lotto/domain/Lotto.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 
 public class Lotto {
@@ -19,5 +20,23 @@ public class Lotto {
         }
     }
 
-    // TODO: 추가 기능 구현
+    public List<Number> getNumbers() {
+        return numbers;
+    }
+
+    private List<Integer> randomNumber() {
+        return Randoms.pickUniqueNumbersInRange(1, 45, 6);
+    }
+
+    private List<Integer> sortNumber(List<Integer> numbers) {
+        return numbers.stream()
+                .sorted()
+                .toList();
+    }
+
+    public Lotto createLotto() {
+        List<Integer> intLottos = randomNumber();
+        List<Integer> sortLottos = sortNumber(intLottos);
+        return new Lotto(sortLottos);
+    }
 }
